Update backport to latest Python3 #30
This updates the backport to CPython 3.7.1+ at this commit:

https://raw.githubusercontent.com/python/cpython/8b9c33ea9ce902f902c9d9900121010801950547/Lib/csv.py

Beyond a few bug fixes, the primary attraction is that the DictReader
now uses OrderedDict ordered mappings.

The tests have been updated accordingly to this commit:
https://raw.githubusercontent.com/python/cpython/8b9c33ea9ce902f902c9d9900121010801950547/Lib/test/test_csv.py

When possible care was taken to minimize the diff output when
comparing against the upstream Python code, therefore a few minor
refactoring are made.

Support for Python 2.6 and Python 3.3 has been dropped.
Python 3.3 cannot run the latest setuptools.
Python 2.6 has no OrderedDict support

The version was bumped to 2.0

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>